core/requirements.txt
asgiref==3.8.1
django==4.2.17
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pip==24.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
setuptools==58.0.4
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1